Consolidate agentFlowEngine
