+++ functions/core/agentFlowEngine.js
@@ -132,4 +132,21 @@ async function runAgentFlow(url, runId, opts = {}) {
   return flowState;
 }
 
-module.exports = { runAgentFlow };
+function createContext({ userId = 'unknown', input = {} }) {
+  return { userId, input, results: {} };
+}
+
+async function runSimpleFlow(flow, { userId = 'unknown', input = {} } = {}) {
+  const ctx = createContext({ userId, input });
+  for (const step of flow) {
+    if (typeof step === 'function') {
+      await step(ctx);
+    }
+  }
+  return ctx.results;
+}
+
+module.exports = {
+  runConfigFlow: runAgentFlow,
+  runAgentFlow: runSimpleFlow
+};
